requirements.txt
asgiref==3.8.1
cfgv==3.4.0
distlib==0.3.8
Django==5.0.7
djangorestframework==3.15.2
filelock==3.15.4
identify==2.6.0
nodeenv==1.9.1
platformdirs==4.2.2
pre-commit==3.8.0
PyYAML==6.0.1
sqlparse==0.5.1
tzdata==2024.1
virtualenv==20.26.3
